Refactor: Replace Java patterns with idiomatic Groovy across codebase
This comprehensive refactoring modernizes the codebase to use idiomatic Groovy
patterns instead of Java-style code, improving readability and maintainability.

Phase 1: Geometry files (GeomHex, GeomStep, GeomPoint, GeomLine, etc.)
- Replaced Math.min/max with .min()/.max() extension methods
- Used .groupBy()/.collectEntries() instead of manual map building
- Simplified null checks with elvis operator

Phase 2: Statistical transformations (GgStat.groovy)
- Replaced Math.min/max/abs with Groovy extensions
- Used .groupBy() for grouping operations

Phase 3: Scale classes (ScaleColorViridis, ScaleFillViridis, etc.)
- Replaced Math.floor/ceil/round with Groovy extensions
- Simplified type casts

Phase 4: Coordinate transformations (CoordPolar, CoordRadial)
- Replaced Math.PI with NumberExtension.PI constant
- Replaced Math.sin/cos/atan2 with Groovy extensions
- Replaced Math.toRadians with .toRadians()

Phase 5: Theme/Guide classes (LegendRenderer, Theme)
- Replaced Math.min with .min()
- Removed unnecessary type casts from list literals

Phase 6: Critical geometry files (GeomDensity2d, GeomContour, GeomBoxplot, etc.)
- Replaced Double.MAX_VALUE with nullable BigDecimal pattern
- Used spread operator (*.getAt()) for collection extraction
- Removed unnecessary toString() calls

Phase 7: Additional geometry files (GeomErrorbar, GeomCrossbar, GeomArea, etc.)
- Verified already using idiomatic patterns

Phase 8: Rendering classes (AxisRenderer, GridRenderer, GgRenderer, etc.)
- Replaced Math.PI, Math.sin/cos with Groovy extensions
- Replaced Math.pow with ** operator for fractional powers
- Fixed type safety issues with intermediate variables

Test file reorganization:
- Moved 62 test files to better align with package structure

All 1545 tests passing after refactoring.

+++ matrix-charts/src/main/groovy/se/alipsa/matrix/gg/aes/CutWidth.groovy
@@ -60,7 +60,7 @@ class CutWidth {
     if (column == null || column.isBlank()) {
       throw new IllegalArgumentException("Column name cannot be null or empty")
     }
-    if (width == null || (width as double) <= 0) {
+    if (width == null || width <= 0) {
       throw new IllegalArgumentException("Width must be a positive number")
     }
     this.column = column
